+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.1.3 - 2.1.3 (????-??-??)
 
+ * 464fd64 - Add PHP templating engine support
  * eb7c407 - Remove trim asset version twig extension & use the service instead
  * 3634c65 - Allow to use custom CKEditor versions
  * cca336a - Extract assets version trim logic in a dedicated service